backend/requirements.txt
Flask
flask-cors
python-dotenv
cohere
PyMuPDF
faiss-cpu
